adds profile update and minor code structure refactor
