+++ backend/app.js
@@ -4,6 +4,7 @@ const cors = require('cors');
 const app = express();
 const db = require('./models');
 const authRouter = require('./routes/authRoutes');
+const profileRouter = require('./routes/profileRoutes');
 const foodItemsRouter = require('./routes/foodItemsRoutes');
 const nearbyRestaurantRouter = require('./routes/nearbyRestaurantRoutes');
 
@@ -15,6 +16,7 @@ app.use(cors({credentials: true, origin: process.env.FRONTEND_URL}));
 
 // Routes
 app.use('/api/auth', authRouter);
+app.use('/api/profile', profileRouter);
 app.use('/api/fooditems', foodItemsRouter);
 app.use('/api/nearbyrestaurants', nearbyRestaurantRouter);
