+++ backend/controllers/authController.js
@@ -137,7 +137,7 @@ const login = async (req, res, next) => {
 
   // Check if username or email and password are provided
   if ((!username && !email) || !password) {
-    return res.status(400).json({
+    return res.status(422).json({
       status: "fail",
       message: "Please provide username or email and password",
     });
@@ -214,57 +214,5 @@ const logout = async (req, res, next) => {
    });
 };
 
-// Profile controller
-const profile = async (req, res, next) => {
-   try {
-      const user = await db.User.findByPk(req.user.id, {
-          attributes: { exclude: ['password'] }
-      });
-
-      if (!user) {
-          return res.status(404).json({
-              status: 'fail',
-              message: 'User not found',
-          });
-      }
-
-      return res.status(200).json({
-          status: 'success',
-          data: user,
-      });
-  } catch (error) {
-      console.error(error);
-      return res.status(500).json({
-          status: 'fail',
-          message: 'Failed to fetch user profile',
-          error: error.message,
-      });
-  }
-};
-
-// Authorization middleware
-const protect = (req, res, next) => {
-   const token = req.cookies.token;
-
-   if (!token) {
-       return res.status(401).json({
-           status: 'fail',
-           message: 'Access denied. No token provided.',
-       });
-   }
-
-   jwt.verify(token, process.env.JWT_SECRET, (err, decoded) => {
-       if (err) {
-           return res.status(401).json({
-               status: 'fail',
-               message: 'Invalid token',
-           });
-       }
-
-       req.user = decoded;
-       next();
-   });
-}
-
-module.exports = { signup, login, profile, logout, protect, verifyToken };
+module.exports = { signup, login, logout, verifyToken };
