+++ backend/routes/foodItemsRoutes.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const { getFoodItems } = require('../controllers/foodItemsController');
-const { protect } = require('../controllers/authController');
+const { protect } = require('../middlewares/authorization');
 const router = express.Router();
 
 router.get('/items', protect, getFoodItems);
