+++ backend/routes/nearbyRestaurantRoutes.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const { getNearbyRestaurants } = require('../controllers/nearbyRestaurantController');
-const { protect } = require('../controllers/authController');
+const { protect } = require('../middlewares/authorization');
 const router = express.Router();
 
 router.get('/', protect, getNearbyRestaurants);
